Removed lowest dependencies tests, use of PSR4
