+++ src/TwbBundle/Options/Factory/ModuleOptionsFactory.php
@@ -16,7 +16,9 @@ class ModuleOptionsFactory implements FactoryInterface
      */
     public function createService(ServiceLocatorInterface $serviceLocator)
     {
-        return $this($serviceLocator, 'TwbBundle\Options\ModuleOptions');
+        $config = $serviceLocator->get('config');
+        $options = $config['twbbundle'];
+        return new ModuleOptions($options);
     }
 
     /**
