+++ tests/TwbBundleTest/ModuleTest.php
@@ -1,5 +1,7 @@
 <?php
+
 namespace TwbBundleTest;
+
 class ModuleTest extends \PHPUnit_Framework_TestCase{
 
 	/**
